change some code for pt4_output.py

+++ pt4_output.py
@@ -14,14 +14,14 @@ class Pt4Out(file):
 
     def pt4getAveCur(self, sec):
         """取得 n 秒后的平均电流"""
-        return "%.2f" % sum(self.curlist100[int(sec)*100:])/(self.curlist100.__len__() - int(sec)*100)
+        return "%.2f" % float(sum(self.curlist100[int(sec)*100:])/(self.curlist100.__len__() - int(sec)*100))
     
     def pt4getAveHighThan(self, curbase):
         """在电流值高过某个值之后就开始计算平均电流"""
         for cur in self.curlist100:
             if cur > int(curbase):
                 pos = self.curlist100.index(cur)
-                return "%.2f" % sum(self.curlist100[pos:])/(self.curlist100.__len__() - pos)
+                return "%.2f" % float(sum(self.curlist100[pos:])/(self.curlist100.__len__() - pos))
         return -1
     
     def pt4getAveLessThan(self, curbase):
@@ -29,7 +29,7 @@ class Pt4Out(file):
         for cur in self.curlist100:
             if cur < int(curbase):
                 pos = self.curlist100.index(cur)
-                return "%.2f" % (sum(self.curlist100[pos:])/(self.curlist100.__len__() - pos))
+                return "%.2f" % float(sum(self.curlist100[pos:])/(self.curlist100.__len__() - pos))
         return -1
     
 if __name__ == "__main__":
